make code more pretty

+++ Commivoyajer_Framework/Models/Output.cs
@@ -3,7 +3,7 @@
     public class Output
     {
         public int[] Sequence { get; set; }
-        public int JourneyLength { get; set; }
+        public double JourneyLength { get; set; }
         public double CalculationTime { get; set; }
         public int variantsCount { get; set; }
     }
